Update week-02-murray-poem.md
Using a quote from the reading.
On page 72, "... anxiety aroused by the fear..."
The tone of the reading is very informal and gives the reader a lot of information, but there are times when the tone changes just with a few words, for example with this quote. The wordplay is simple but effective. Using tone and playing with the viewer's inner voice while reading this chapter. Then the tone is dark and works up the fun beat, but the way I read it was an upbeat tone in the beginning then it got exciting. Finally, the tone is neutral like it's waiting for the reader to finish the the story of the moving picture.