Update week-02-murray-poem.md
In line 2, I wanted to summarize the constant evolution of wanting to create more, but the fear and the "fog" when symbolizing human fear of the unknown. I used parts of a camera like the lens being the eyes and the film breathing in a way. Like if something is being created here. The thing being created is the Internet.