Update week-02-murray-poem.md
The loop is how even with the evolution of the web there would so be information from the past. Whether it be from us or with fear. Like in the quote that I have used, symbolizing the fear that most people; are being judged and made up of, but in a way we still use the intent. It is a constant loop of anxiety for some.