Update week-02-murray-poem.md
The fear that people had with the intent has changed in a way where the intent there are endless options, but like with free will. That endless freedom can consume its viewers.

I used the quote and made a smoother transition to make the flow of the poem more smoother.